+++ Studio/InstancesModel.cpp
@@ -47,7 +47,7 @@ void InstancesModel::updateInternal(Audio::BeatRanges *data)
 
 void InstancesModel::add(const BeatRange &range)
 {
-    const auto idx = std::distance(_data->begin(), _data->findSortedPlacement(range));
+    const auto idx = static_cast<int>(std::distance(_data->begin(), _data->findSortedPlacement(range)));
 
     Scheduler::Get()->addEvent(
         [this, range] {
@@ -99,7 +99,7 @@ const BeatRange &InstancesModel::get(const int idx) const noexcept_ndebug
 
 void InstancesModel::set(const int idx, const BeatRange &range)
 {
-    auto newIdx = std::distance(_data->begin(), _data->findSortedPlacement(range));
+    auto newIdx = static_cast<int>(std::distance(_data->begin(), _data->findSortedPlacement(range)));
 
     coreAssert(idx >= 0 && idx < count(),
         throw std::range_error("InstancesModel::move: Given index is not in range: " + std::to_string(idx) + " out of [0, " + std::to_string(count()) + "["));
